Gazebo timer node controls gazebo's max update rate to ensure sensor fps
Gazebo controls its own update rate, yet sometimes (especially in docker
containers) sensors are not updated often enough.
This commit introduces the gazebo_timer_node which reduces the
max_update_rate of gazebo such that the sensors are updated as expected.
The control of the gazebo update rate is very rudimentary!