NodeBase: set active variable in run function instead of start/stop
Setting the __active variable in the start/stop function made the behavior of the node dependent on when the super().start()/stop() was called by child classes.
E.g. Shutting down publishers before calling super().stop() resulted in errors!
This is solved by moving all functionality into the run() function.

(Also the Rate class is used to make the code a little nicer)