Create StateMachineNode
This commit introduces the StateMachineNode which tracks the state of
the current drive. There are four state machines with each tracking only
a part. The four state machines are: OvertakingStateMachine,
ParkingStateMachine, PriorityStateMachine and ProgressStateMachine.
Each StateMachine and the Node itself comes with a (ROS)Test.

The Node subscribers to the Speaker and publishes whener one
StateMachine changes its state.