Create GroundtruthNode with a mock, visualization and ROS test
The GroundtruthNode creates a connection between the road's groundtruth and ROS. This commit additionally comes with a GroundtruthMockNode which advertises a mocked road which is useful for testing. Additionally a visualization allows to visualize the groundtruth in RVIZ.
There's also a ROS test for the GroundtruthNode and an extensive documentation.