Create SpeakerNode and EvaluationTestNode
This commit introduces the SpeakerNode, a main component of the simulation evaluation pipeline.  It can be used to understand processes happening in the simulated world (e.g. car driving on the right side of the road). The SpeakerNode creates a number of speakers which all deal with separate tasks of understanding the simulated world.
The speakers and the SpeakerNode are tested with unittest and a ROS test.

There's also the EvaluationTestNode which allows to easily test the behavior of the SpeakerNode and in the following also subsequent parts of the evaluation pipeline.