Generate dr drift model within gazebo_simulation using the urdf package
The car model can now be generated using
```
rosrun gazebo_simulation generate_dr_drift
```
Resulting files are used as parameters when launching the simulation. The vehicle is spawned on launch.

The car_model sub directory in gazebo simulation processes information about the real car's sensors and specifications to produces a model that is as close to reality as possible.