Ignore undefined fields in Car/CameraSpecs, unify load/save in Specs
Whenever a field was added to car/camera specs in kitcar ros, the :py:func`from_file` would try to pass that field to the __init__ of CarSpecs, which failed if that field was not manually added to CarSpecs as well.
Now only fileds that are defined within CarSpecs are passed to __init__.
The save/ load methods are unified in a common super class called Specs.