Prevent vehicle_simulation_link_node test to be blocked waiting for tf
The vehicle_simulation_link_node's test waits until it recteives a /tf message to signal that the vehicle_simulation_link_node is started.
At the same time the node waits for a state_estimation message that is published by the test.
This counter dependency could lead to the node;s waiting for each other.
These changes prevent that by publishing state machine messages continuously until a response is received.

+ some cleaning up