Ensure transforms are broadcast at most once per time stamp
WIth ROS noetic a warning is raised whenever a transform is published multiple times for a single time stamp.
The issue happens within the simulation which might be due to the discrete update rate of 1000 Hz.

There is probably a deeper reason why the duplicate transforms occur in the first place but I could not find any reason yet.