Add default image+kitcar_ros_ci image, rename CI image
The previously available ci_image is now simply called ci. It contains all packages necessary to build the simulation and is used as a basis within the Gitlab ci.
The default image is a docker image that contains a ready to use version of the simulation. It builds on top of the ci image.
The kitcar_ros_ci image builds on top of the default image but also contains packages necessary to build kitcar-ros. It can therefore be used to run the simulation in the kitcar-ros CI.

An additional docs page is meant to explain the intentions and usage of the docker images.