Fix automatic drive test and extract path into separate param file
Within the test_automatic_drive file, the current position projected onto the line was subtracted by the lines length and the test was successful once this difference was smaller than some constant.
This was always the case, because the abs() was not taken!

Extracting the path parameter file into it's own file allows to easily start the automatic_drive.launch file with a different path.