init: Separate necessary pip packages from additional ones
The init script now asks if packages only required to build the documentation or to run machine learning tasks should be installed.
These packages are large and may require a lot of time to install. Making the installation optional reduces the size of docker images and saves people some time.