cycle_gan: Add optional cycle noise during training
During training the generators tend to hide information within the generated images for the other generator to decode and better regenerate the original image. Thus reducing the cycle loss.
We don't want this behavior as it limits the generators capability to produce real looking images because the cycle loss will increase once the generated image does not contain the hidden information any more.
The introduced gaussian noise helps to limit the generators capability to hide information from the discriminator, according to https://arxiv.org/pdf/1908.01517.pdf.