camera: Rewrite node in python & no more optional noise added to image
The camera node is now written in python to allow easier integration of other features, like the cycle gan in the future.
This requires the cv bridge to work with python3.

The apply_gan parameter of the master.launch file now allows to transform camera input images using a trained generator of the cycle gan.