ci: Fix blocking pipeline job in merge requests
The deploy-docker-images Job should be optional for non-master branches but was causing a blocked pipeline in merge request and thus had to be executed every time.

I have used only/except instead of rules as it seems to have a bug that causes the detached pipelines. See: https://gitlab.com/gitlab-org/gitlab/-/issues/34756