gazebo_rate_control: Fix waiting for non existing topic
If a topic was defined as a target of the gazebo rate control and was not published at all, the measured rate was 0 for that topic and thus too low. The adjusted max_update_rate would be reduced as far as possible.
Now topics that are not published at all, are disregarded by the node.