road: Incorporate transformable&fix transform to parking lots
Subclassing RoadSection and RoadElement from Transformable unifies how transforms to the objects are defined.
The set_transform method is used to update all objects that must be updated when the transform changes.

Additionally, the ParkingLot is slightly changed. Until now the origin of a parking area was used as the origin of it's parking lots. Now, each parking lot has its own origin at the lower left corner of the left most spot, when looking from the center of the road to the individual side of the road (left/right).