camera&image_translator: Preserve part of original image
Optionally the gan can be used to only overlay the original image and modify it slightly instead of completely changing it. Two parameters can be used to variably adjust how much of the original image is kept.

Colored pixels are used to prevent the GAN from changing traffic signs so much that they are no longer recognizable.